2022-07-25 JSP&Servlet 삭제, 검색기능

+++ StudyJspServlet/.metadata/.plugins/org.eclipse.wst.server.core/tmp0/wtpwebapps/05.Project/notice/detail.jsp
@@ -35,7 +35,7 @@
 	<!-- 작성자로 로그인한 경우만 수정/삭제 가능 -->
 	<c:if test='${userInfo.userid eq dto.writer}'>
 	<a class='btn-fill' href='modify.no?id=${dto.id}'>정보수정</a>
-	<a class='btn-fill' href='delete.no?id=${dto.id}'>정보삭제</a>
+	<a class='btn-fill' onclick="if( confirm('정말 삭제?') ) href='delete.no?id=${dto.id}' ">정보삭제</a>
 	</c:if>
 	<!-- 로그인된 경우 답글쓰기 가능 -->
 	<c:if test='${!empty userInfo}'>
